pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "picpocket"
version = "0.1.0-dev"
authors = [
  { name="bcj", email="bcj@bcjbcj.com" },
]
description = "A place to store photos"
readme = "README.md"
# license = {file = "LICENSE.md"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    "Environment :: Console",
    # "Environment :: MacOS X",
    # "Environment :: No Input/Output (Daemon)",
    # "Environment :: Web Environment",
    # "Intended Audience :: End Users/Desktop",
    "License :: Other/Proprietary License",
    # "Operating System :: MacOS",
    # "Operating System :: Microsoft",
    # "Operating System :: Microsoft :: Windows",
    # "Operating System :: Microsoft :: Windows :: Windows 10",
    # "Operating System :: Microsoft :: Windows :: Windows 11",
    # "Operating System :: POSIX",
    # "Programming Language :: JavaScript",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: SQL",
    "Topic :: Database",
    # "Topic :: Desktop Environment",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    # "Topic :: Multimedia",
    # "Topic :: Multimedia :: Graphics :: Viewers",
    # "Topic :: System",
    # "Topic :: System :: Archiving",
    # "Topic :: System :: Archiving :: Backup",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "aiosqlite",
    "cryptography",
    "pillow",
    "tomli-w",
    "tornado",
]

[project.optional-dependencies]
PostgreSQL = ["psycopg"]

[project.urls]
"Homepage" = "https://github.com/bcj/picpocket-py"

[project.scripts]
picpocket = "picpocket.cli:with_default"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]